fix: avoid virtual list jumps to the end on new message from another author

+++ src/dotnet/UI.Blazor/Components/VirtualList/VirtualListDataQuery.cs
@@ -6,8 +6,8 @@ public class VirtualListDataQuery
 
     public Range<string> KeyRange { get; }
     public Range<double>? VirtualRange { get; init; }
-    public double ExpandStartBy { get; init; } = 0;
-    public double ExpandEndBy { get; init; } = 0;
+    public int ExpandStartBy { get; init; } = 0;
+    public int ExpandEndBy { get; init; } = 0;
 
     public bool IsNone
         => ReferenceEquals(this, None);
